+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 section below it for the last release. -->
 ## \<next>
 
+* The configured project is now automatically reloaded when adding/renaming/deleting a file in a known source directory.
+* GPR .cgpr config files can now be passed to the Ada Language Server through the new `gprConfigurationFile` configuration option.
+* Project diagnostics are no longer published on each opened Ada file: instead, they are now published on the workspace's root directory.
 * Add a setting `ada.showNotificationsOnErrors` to control whether VS Code should display error notifications for failing LSP requests.
 * Add support for a default `.als.json` workspace configuration file and a global `$XDG_CONFIG_HOME/als/config.json` user configuration file. See [settings.md](https://github.com/AdaCore/ada_language_server/blob/master/doc/settings.md) for more information.